requirements.txt
alembic==1.8.1
click==8.1.3
rich==12.5.1
discord.py==2.0.1
psycopg2==2.9.3
gino==1.0.1
